test/test-functions: use virtio for testbed drive
For qemu tests on x86, it's ok to expect the disk driver to work without
any special config and the disk naming to be /dev/sda.  However on other
archs, disk drivers do not always work without additional arch-specific
config, and the disk name is not always /dev/sda.

This changes qemu to use the virtio disk driver, which is supported
under all archs without any additional config needed, as well as always
using the disk naming of /dev/vda.