test/test-functions: expand number of archs to look for qemu-system-*
Increase the number of archs where a qemu-system-* binary is searched for;
currently qemu is searched for on x86_64, i*86, and ppc64*.  This adds
arm* and aarch64*, and falls back to qemu-system-$ARCH for all other
archs.

The x86_64 specific case is removed, since the default fallback to
qemu-system-$ARCH works for that arch.

For backward compatibility, if the arch-specific binary name isn't
found, the plain "qemu" binary name is also checked.