watchdog: add support for fifo watchdog
This patch adds a fifo watchdog to systemd.

At runtime systemd will detect the watchdog device mode based on
watchdog file type. If the the configured file is a device node the
watchdog is handled as before using device driver interface for hardware
watchdog. If the file is a fifo, the protocol for handlling watchdog in
this mode is applied.

The watchdog on fifo protocol follows the same steps as hardware watchdog
but writing ascii text messages on the fifo instead of doing ioctl() calls.
Each message starts with MACHINE and HOSTNAME fields (sd-id128 and
machine hostname) followed by the command and value (see watchdog_set_message()).
Commands are send for setting the timeout (SET_TIMEOUT), send heartbeat
(KEEPALIVE) and disarm the watchdog (DISARM).

This enables using HW watchdog on a native host with systemd and fifo based
watchdog to a container running its own systemd init.