networkd: drop DHCPv4 config on stop, keep on restart, by default
This partially reverts 95355a281c06c5970b7355c38b066910c3be4958.

It seems that other parts of link_stop_clients() should be skipped
when restarting, but I don't know enough about those other clients to have
an opinion if it is better to stop&start them on restart or not.
Anyway, that can be done in later patches now that the support for restarts
is there.

Fixes #13625.