dbus-unit: Make Slice= live property
If a DBus caller intends to run a non-transient unit under a specified
slice, it cannot change the Slice= without creating a drop-in and
complete daemon-reload. Let us allow changing Slice= dynamically via
DBus (convert it from transient property to live property).

Note it is not a real live property, we want to avoid migrating a
running unit under a different slice. Hence, rely on the checks in
unit_set_slice() to make sure we are only changing Slice= of a unit that
is not running currently.