Add systemd-tmpfiles-setup-var-log-journal.service
In case /var/log/journal is on a volume that depends on the network
for any reason, ordering systemd-journal-flush before
systemd-tmpfiles-setup creates a cycle that breaks booting, as
systemd-tmpfiles-setup is a dependency of systemd-networkd, which
means mounting /var/log/journal is ordered to happen both before and
after the network is online.

Add a new tmpfiles.d and a new unit specifically for /var/log/journal,
so that the dependency cycle can be broken.

Fixes #13714