core: Run FailureAction even when a unit hasn't started
When a unit enters the failed state, it will start the OnFailure= unit.
If the unit was previously running, it will also run its FailureAction.

This is unhelpful if for example you have a have a unit designed to
reboot if a subset of system services it depends on fail.