core, job: handle reload as JOB_TRY_RELOAD, not JOB_RELOAD_OR_START
Currently, if the target service unit is in the activating state, systemctl
reload command now waits for the target service to finish the starting. This
behavior is odd. The service unit in the activating state is the service not
running, reload operation should be considered as NOP.

The cause of this implementation is from a little confusion about
JOB_RELOAD_OR_START. This is not for reload command. JOB_RELOAD_OR_START is for
reload-or-restart command, which includes restart command and starts a given
service unit if it's not running. The job type we should use for reload command
is JOB_TRY_RELOAD. In other words, reload command is inherently try-reload
command.

The change takes advantage of the existing code for JOB_TRY_RELOAD and
JOB_TRY_RESTART, thanks to which, it's relatively small.

By this change, job_type_lookup_merge() never receives a combination of
(JOB_RELOAD, JOB_START) because if there's start job then the job type is
collapsed to JOB_NOP. Hence, replace the value of the corresponding entry by
-1.

Related: #10464